move some functions to utils.py
